feat: 优化 Homepage 架构并启用 Blog ISR
重构 Homepage 组件结构，提升代码可维护性，并为 Blog 页面添加增量静态再生成（ISR）以提升性能。

主要改动：
- 将 Homepage 组件模块化到 app/_components/homepage/ 目录
- 新增独立组件：Header, Footer, Hero, Features, Contact 等
- Blog 详情页添加 ISR（每小时重新生成）和 SSG 预生成（前 50 篇）
- 启用 Turbopack 构建以提升编译速度
- 优化 Next.js 配置：日志输出、图片加载、性能监控
- 新增分享页面布局文件
- 优化 Blog API 类型定义（支持 limit 和 page 参数）

性能提升：
- Homepage 代码体积减少 500+ 行
- Blog 页面响应速度提升（ISR + SSG）
- 开发和生产构建速度加快（Turbopack）
