+++ src/app/_components/homepage/index.ts
@@ -1,0 +1,6 @@
+export { Header } from './header';
+export { Hero } from './hero';
+export { Features } from './features';
+export { Contact } from './contact';
+export { Footer } from './footer';
+export { SITE_CONFIG, STRUCTURED_DATA } from './constants';
